+++ build.gradle.kts
@@ -2,4 +2,5 @@ plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.org.jetbrains.kotlin.android) apply false
     alias(libs.plugins.android.library) apply false
+    alias(libs.plugins.jetbrains.kotlin.jvm) apply false
 }
