+++ gradle/libs.versions.toml
@@ -6,10 +6,12 @@ androidx_test_ui_automator = "2.3.0"
 appcompat = "1.7.0"
 constraintlayout = "2.1.4"
 coreKtx = "1.13.1"
+dokka = "1.9.20"
 espressoCore = "3.6.1"
 junitVersion = "1.2.1"
 jupiter = "5.11.0"
 kotlin = "2.0.10"
+kotlinxCoroutinesCore = "1.7.3"
 logback-android = "3.0.0"
 logback-classic = "1.5.6"
 material = "1.12.0"
@@ -18,6 +20,7 @@ navigationUiKtx = "2.7.7"
 slf4jApi = "2.0.16"
 mannodermaus_junit5_plugin = "1.11.0.0"
 mannodermaus_junit5 = "1.5.0"
+jetbrainsKotlinJvm = "2.0.10"
 
 [libraries]
 androidx-appcompat = { module = "androidx.appcompat:appcompat", version.ref = "appcompat" }
@@ -33,6 +36,7 @@ de-manodermaus-android-junit5 = { module = "de.mannodermaus.junit5:android-test-
 de-manodermaus-android-junit5-runner = { module = "de.mannodermaus.junit5:android-test-runner", version.ref = "mannodermaus_junit5" }
 junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version.ref = "jupiter" }
 junit-jupiter-api = { module = "org.junit.jupiter:junit-jupiter-api", version.ref = "jupiter" }
+kotlinx-coroutines-core = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "kotlinxCoroutinesCore" }
 logback-android = { module = "com.github.tony19:logback-android", version.ref = "logback-android" }
 logback-classic = { module = "ch.qos.logback:logback-classic", version.ref = "logback-classic" }
 material = { module = "com.google.android.material:material", version.ref = "material" }
@@ -42,7 +46,7 @@ packetdumper = { module = "com.jasonernst.packetdumper:packetdumper", version = 
 
 [bundles]
 android-test = ["junit-jupiter-api", "androidx_test_runner", "androidx_test_monitor", "androidx_test_uiautomator", "de-manodermaus-android-junit5"]
-unit-test = ["junit-jupiter-api", "de-manodermaus-android-junit5"]
+unit-test = ["junit-jupiter-api", "junit-jupiter"]
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
@@ -52,4 +56,6 @@ git-version = { id = "me.qoomon.git-versioning", version = "6.4.3" }
 sonatype-maven-central = { id = "com.vanniktech.maven.publish", version = "0.29.0" }
 gradleup-nmcp = { id = "com.gradleup.nmcp", version = "0.0.9"}
 de-mannodermaus-android-junit5 = { id = "de.mannodermaus.android-junit5", version.ref = "mannodermaus_junit5_plugin" }
+dokka = { id = "org.jetbrains.dokka", version.ref = "dokka" }
 cmake = { id = "dev.welbyseely.gradle-cmake-plugin", version = "0.1.0"}
+jetbrains-kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "jetbrainsKotlinJvm" }
