+++ icmp-android/build.gradle.kts
@@ -1,0 +1,104 @@
+plugins {
+    alias(libs.plugins.org.jetbrains.kotlin.android)
+    alias(libs.plugins.de.mannodermaus.android.junit5)
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.git.version) // https://stackoverflow.com/a/71212144
+    alias(libs.plugins.sonatype.maven.central)
+    alias(libs.plugins.gradleup.nmcp)
+    alias(libs.plugins.dokka)
+    id("signing") // https://medium.com/nerd-for-tech/oh-no-another-publishing-android-artifacts-to-maven-central-guide-9d7f300ebd74
+}
+
+android {
+    namespace = "com.jasonernst.icmp_lib"
+    compileSdk = 34
+
+    defaultConfig {
+        minSdk = 29
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
+    }
+    packaging {
+        resources {
+            excludes.add("/META-INF/{AL2.0,LGPL2.1}")
+            excludes.add("/META-INF/INDEX.LIST")
+        }
+    }
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+        }
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = "17"
+    }
+}
+
+dependencies {
+    implementation(project(":icmp-common"))
+    implementation(libs.logback.android)
+
+    androidTestImplementation(libs.bundles.android.test)
+    androidTestRuntimeOnly(libs.de.manodermaus.android.junit5.runner)
+}
+
+version = "0.0.0-SNAPSHOT"
+gitVersioning.apply {
+    refs {
+        branch(".+") { version = "\${ref}-SNAPSHOT" }
+        tag("v(?<version>.*)") { version = "\${ref.version}" }
+    }
+}
+
+// see: https://github.com/vanniktech/gradle-maven-publish-plugin/issues/747#issuecomment-2066762725
+// and: https://github.com/GradleUp/nmcp
+nmcp {
+    val props = project.properties
+    publishAllPublications {
+        username = props["centralPortalToken"] as String? ?: ""
+        password = props["centralPortalPassword"] as String? ?: ""
+        // or if you want to publish automatically
+        publicationType = "AUTOMATIC"
+    }
+}
+
+// see: https://vanniktech.github.io/gradle-maven-publish-plugin/central/#configuring-the-pom
+mavenPublishing {
+    coordinates("com.jasonernst.icmp_lib", "icmp_android", version.toString())
+    pom {
+        name = "ICMP Android"
+        description = "A library for sending and receiving ICMP packets on Android"
+        inceptionYear = "2024"
+        url = "https://github.com/compscidr/icmp"
+        licenses {
+            license {
+                name = "GPL-3.0"
+                url = "https://www.gnu.org/licenses/gpl-3.0.en.html"
+                distribution = "repo"
+            }
+        }
+        developers {
+            developer {
+                id = "compscidr"
+                name = "Jason Ernst"
+                url = "https://www.jasonernst.com"
+            }
+        }
+        scm {
+            url = "https://github.com/compscidr/icmp"
+            connection = "scm:git:git://github.com/compscidr/icmp.git"
+            developerConnection = "scm:git:ssh://git@github.com/compscidr/icmp.git"
+        }
+    }
+
+    signAllPublications()
+}
