+++ icmp-android/src/main/AndroidManifest.xml
