+++ icmp-common/src/main/java/com/jasonernst/icmp_common/PacketHeaderException.kt
@@ -1,0 +1,32 @@
+/*
+ *  Copyright 2014 AT&T
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+package com.jasonernst.icmp_common
+
+import java.lang.Exception
+
+/**
+ *
+ * @author Borey Sao
+ * Date: May 7, 2014
+ */
+class PacketHeaderException : Exception {
+    constructor(message: String?) : super(message) {}
+    constructor(message: String?, throwable: Throwable?) : super(message, throwable) {}
+
+    companion object {
+        private const val serialVersionUID = 1L
+    }
+}
